+++ utils/spx_getconf.js
@@ -3,7 +3,6 @@
 // Note, these functions execute BEFORE logger or utils are loaded
 // so none of those functions can be used here!
 
-
 const fs = require('fs');
 const path = require('path');
 const { rejects } = require('assert');
@@ -25,7 +24,7 @@ module.exports = {
             }
         catch (error)
             {
-                console.log('makeFolderIfNotExist: error while checking or creatuing folder [' + fullfolderpath + '].' + error);
+                console.log('makeFolderIfNotExist: error while checking or creating folder [' + fullfolderpath + '].' + error);
                 return false
             }
       },
@@ -35,23 +34,20 @@ module.exports = {
         // read config.json. Usage: "cfg.readConfig()"
         return new Promise(resolve => {
             try {
-                let CONFIG_FILE = path.join(process.cwd(), 'config.json'); // tested on ubuntu
-                // console.log('Trying ' + CONFIG_FILE);
-                // console.log('Command line arguments given: [' + process.argv + '].');
 
+                let CONFIG_FILE = path.join(process.cwd(), 'config.json'); // tested on ubuntu
                 // check 1st commandline argument (after node and script, that is)
                 var myArgs = process.argv.slice(2);
                 var ConfigArg = myArgs[0] || '';
                 if (ConfigArg){
-                    // console.log('Command line arguments given: [' + process.argv + ']. Reading config from ' + ConfigArg + '.');
-                    console.log('Using configfile [' + ConfigArg + '].');
+                    console.log('Command line arguments given: [' + process.argv + '], reading config from ' + ConfigArg + '.');
                     CONFIG_FILE = path.join(process.cwd(), ConfigArg); // tested on ubuntu    
                 }
                 
                 // check if config file exists
                 if (!fs.existsSync(CONFIG_FILE)) {
                     // config file not found, let's create the default one
-                    console.log('Specified config file (' + CONFIG_FILE + ') not found, generating defaults...');
+                    console.log('** Config file (' + CONFIG_FILE + ') not found, generating defaults. **');
                     let cfg                                     = {}
                     cfg.general                                 = {}
                     cfg.general.username                        = "welcome"
@@ -61,9 +57,10 @@ module.exports = {
                     cfg.general.loglevel                        = "info"
 
                     // below paths were __dirname but pkg did not like it
-                    cfg.general.logfolder                       = path.join(process.cwd(), 'LOG').replace(/\\/g, "/") + "/";
-                    cfg.general.dataroot                        = path.join(process.cwd(), 'DATAROOT').replace(/\\/g, "/") + "/";
-                    cfg.general.templatefolder                  = path.join(process.cwd(), 'ASSETS/templates').replace(/\\/g, "/")+ "/";
+                    cfg.general.logfolder                       = path.join(process.cwd(), 'LOG').replace(/\\/g, "/") + "/"
+                    cfg.general.dataroot                        = path.join(process.cwd(), 'DATAROOT').replace(/\\/g, "/") + "/"
+                    cfg.general.templatefolder                  = path.join(process.cwd(), 'ASSETS/templates').replace(/\\/g, "/")+ "/"
+                    cfg.general.templatesource                  = "spxgc-ip-address"
 
                     cfg.general.port                            = "5000"
                     cfg.casparcg                                = {}
@@ -104,16 +101,16 @@ module.exports = {
                     newcontrol3.description                      = "Clear playout channels"
                     cfg.globalExtras.CustomControls.push(newcontrol3)
 
-
                     // Write config file. Note, this does not use utility function.
                     cfg.warning = "GENERATED DEFAULT CONFIG. Modifications done in the GC will overwrite this file.";
-                    cfg.smartpx = "(c) 2020 Tuomo Kulomaa <tuomo@smartpx.fi>";
+                    cfg.smartpx = "(c) 2020-2021 Tuomo Kulomaa <tuomo@smartpx.fi>";
                     cfg.updated = new Date().toISOString();
                     global.config = cfg; // <---- config to global scope
                     let filedata = JSON.stringify(cfg, null, 2);
-                    console.log("Writing " + CONFIG_FILE)
+                    console.log('** Writing default config values to [' + CONFIG_FILE + '] **\n') 
                     fs.writeFileSync(CONFIG_FILE, filedata, 'utf8', function (err) {
                     if (err){
+                        console.error("Error writing default config to [" + CONFIG_FILE + "]")
                         process.exit(2)
                         throw error;
                         }
@@ -121,15 +118,14 @@ module.exports = {
                 }
                 else {
                     // file was found, let's read and use that
+                    // console.log("Loading config from " + CONFIG_FILE + "...")
                     let configFileStr = fs.readFileSync(CONFIG_FILE);
                     global.config = JSON.parse(configFileStr);
                 }
 
-
                 // folderchecks
                 this.makeFolderIfNotExist(global.config.general.logfolder);
                 this.makeFolderIfNotExist(global.config.general.dataroot);
-
                 global.configfileref =  CONFIG_FILE; // this is it!
                 resolve()
             }